+++ DLLServices2TestStand/Properties/Resources.Designer.cs
@@ -19,7 +19,7 @@ namespace DLLServices2TestStand.Properties {
     // class via a tool like ResGen or Visual Studio.
     // To add or remove a member, edit your .ResX file then rerun ResGen
     // with the /str option, or rebuild your VS project.
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "16.0.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "17.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     public class Resources {
@@ -84,6 +84,15 @@ namespace DLLServices2TestStand.Properties {
         public static string CMDARG_BASE64_ENCODING {
             get {
                 return ResourceManager.GetString("CMDARG_BASE64_ENCODING", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to CaseInsensitiveStringReplace.
+        /// </summary>
+        public static string CMDARG_CASE_INSENSITIVE_STRING_REPLACE {
+            get {
+                return ResourceManager.GetString("CMDARG_CASE_INSENSITIVE_STRING_REPLACE", resourceCulture);
             }
         }